+++ x/crosschain/keeper/migrator.go
@@ -6,6 +6,7 @@ import (
 	v3 "github.com/zeta-chain/zetacore/x/crosschain/migrations/v3"
 	v4 "github.com/zeta-chain/zetacore/x/crosschain/migrations/v4"
 	v5 "github.com/zeta-chain/zetacore/x/crosschain/migrations/v5"
+	v6 "github.com/zeta-chain/zetacore/x/crosschain/migrations/v6"
 )
 
 // Migrator is a struct for handling in-place store migrations.
@@ -38,4 +39,8 @@ func (m Migrator) Migrate3to4(ctx sdk.Context) error {
 // Migrate4to5 migrates the store from consensus version 4 to 5
 func (m Migrator) Migrate4to5(ctx sdk.Context) error {
 	return v5.MigrateStore(ctx, m.crossChainKeeper, m.crossChainKeeper.zetaObserverKeeper)
+}
+
+func (m Migrator) Migrate5to6(ctx sdk.Context) error {
+	return v6.MigrateStore(ctx, m.crossChainKeeper)
 }