+++ x/observer/client/cli/tx_updat_observer_test.go
@@ -1,0 +1,13 @@
+package cli_test
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/zeta-chain/zetacore/x/observer/client/cli"
+)
+
+func Test_UpdateObserver(t *testing.T) {
+	r, err := cli.ParseUpdateReason(0)
+	fmt.Println(r, err)
+}