+++ x/observer/client/cli/tx_update_observer.go
@@ -28,7 +28,7 @@ func CmdUpdateObserver() *cobra.Command {
 				return err
 			}
 			// #nosec G701 parsed in range
-			updateReason, err := parseUpdateReason(int32(updateReasonInt))
+			updateReason, err := ParseUpdateReason(int32(updateReasonInt))
 			if err != nil {
 				return err
 			}
@@ -48,7 +48,7 @@ func CmdUpdateObserver() *cobra.Command {
 	return cmd
 }
 
-func parseUpdateReason(i int32) (types.ObserverUpdateReason, error) {
+func ParseUpdateReason(i int32) (types.ObserverUpdateReason, error) {
 	if _, ok := types.ObserverUpdateReason_name[i]; ok {
 		switch i {
 		case 1: